Download model at application startup time
The slug size on Heroku is limited to 500 MB. Thus, large models cannot
be contained in the slug that is produced by the Heroku build process.

This means that, if we have a large model:

- We cannot include it in the Git repository that is pushed to
  Heroku (since the model would end up in the slug)
- We cannot download it during the Heroku build (e.g. with a custom
  buildpack), since in this case the model would also end up in the slug

Therefore, we download and install the model at application startup
time. This makes each startup slower, but it seems to be the only way to
work around the slug size limitation of 500 MB imposed by Heroku.