Integrate with CloudAMQP and chatbot
The main class now just starts the RPC server.

The RPC server accepts requests from the chatbot, parses the sentences
contained in the requests with the EasyCCG parser, and sends the resulting
parse trees back to the chatbot.

At the moment, the requests must consist of a single sentence as a
string. The server returns only the first parse tree of this sentence to
the chatbot.